requirements.txt
streamlit==1.37.1
pillow==10.4.0
numpy==1.26.4
pandas==2.2.2
openpyxl==3.1.5
opencv-python<=4.6.0.66

# PaddleOCR (CPU)
paddlepaddle==2.6.1
paddleocr==2.7.3
